test(server): split auth and route surface lib tests
