+++ nodelite-server/src/lib_tests.rs
@@ -5,8 +5,7 @@ use std::time::{SystemTime, UNIX_EPOCH};
 
 use axum::Router;
 use axum::body::{Body, to_bytes};
-use axum::extract::{ConnectInfo, State};
-use axum::http::{HeaderMap, HeaderValue, Request, StatusCode, header};
+use axum::http::{HeaderMap, Request, StatusCode, header};
 use axum::middleware::{from_fn, from_fn_with_state};
 use axum::routing::{get, post};
 use chrono::Utc;
@@ -14,9 +13,7 @@ use serde_json::json;
 use tokio::runtime::Runtime;
 use tower::util::ServiceExt;
 
-use super::{
-    AppState, ServerReadiness, set_protected_response_headers, uses_insecure_remote_public_base_url,
-};
+use super::{AppState, set_protected_response_headers};
 use super::support::{
     assert_security_headers, json_request, json_write_routes, protected_ok, protected_request,
     small_json_write_requests, trusted_proxies, two_factor_auth_test_state, ws_upgrade_request,
@@ -26,407 +23,17 @@ use crate::admission::{
     resolve_client_ip, sweep_expired_auth_failures,
 };
 use crate::audit::{AuditEvent, AuditEventType, AuditQuery, NewAuditEvent};
-use crate::auth::{ReadonlyRouteAuth, TwoFactorSessions};
 use crate::handlers::{
-    audit_log, bootstrap, healthz, index, install_agent_script, install_bootstrap,
-    is_well_formed_install_token, logout_and_reauth, node_detail, node_history, node_logs,
-    node_status, nodes, overview, readyz, require_readonly_auth, static_asset, verify_2fa_page,
+    audit_log, index, is_well_formed_install_token, logout_and_reauth, require_readonly_auth,
+    verify_2fa_page,
 };
-use crate::registry::{IssueNodeRequest, issue_node};
 use crate::sanitize::{
     MAX_SANITIZED_DISKS, MAX_SANITIZED_LOAD, MAX_SANITIZED_RATE_BYTES_PER_SEC,
     MAX_SANITIZED_STRING_BYTES, METRIC_ANOMALY_SESSION_LIMIT, SanitizationReport,
     sanitize_snapshot, should_disconnect_for_metric_anomalies, update_metric_anomaly_window,
 };
 use crate::test_support::{TEST_BASIC_AUTH_HEADER, test_server_config, test_ws_config};
-use crate::ws::ws_handler;
 use nodelite_proto::{NodeSnapshot, ServerConfig, WsConfig};
-use tower_http::trace::TraceLayer;
-
-#[test]
-fn router_builds_with_v08_path_syntax() {
-    let unique = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .expect("clock should be monotonic enough")
-        .as_nanos();
-    let registry_path = std::env::temp_dir().join(format!("nodelite-router-test-{unique}.json"));
-    let mut config = test_server_config(
-        SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 8080)),
-        "http://127.0.0.1:8080".to_string(),
-        registry_path,
-        PathBuf::from("./data/history.sqlite3"),
-        PathBuf::from("./data/snapshot.json"),
-    );
-    config.readonly_auth = None;
-    config.ws = test_ws_config(32, 8);
-    config.stale_after_secs = 20;
-    config.ping_interval_secs = 10;
-    config.ignored_filesystems = vec!["tmpfs".to_string()];
-    let config = Arc::new(config);
-    let runtime = Runtime::new().expect("runtime should build");
-    let state = runtime
-        .block_on(AppState::test_fixture(
-            config,
-            Arc::new(PathBuf::from("config/server.toml")),
-        ))
-        .expect("state fixture should build");
-
-    let _app: Router = Router::new()
-        .route("/", get(index))
-        .route("/nodes/{node_id}", get(node_detail))
-        .route("/assets/{*path}", get(static_asset))
-        .route("/healthz", get(healthz))
-        .route("/readyz", get(readyz))
-        .route("/install/install-agent.sh", get(install_agent_script))
-        .route("/install/bootstrap", get(install_bootstrap))
-        .route("/api/bootstrap", get(bootstrap))
-        .route("/api/overview", get(overview))
-        .route("/api/nodes", get(nodes))
-        .route("/api/audit-log", get(audit_log))
-        .route("/api/nodes/{node_id}", get(node_status))
-        .route("/api/nodes/{node_id}/history", get(node_history))
-        .route("/api/nodes/{node_id}/logs", get(node_logs))
-        .route("/ws", get(ws_handler))
-        .with_state(state)
-        .layer(TraceLayer::new_for_http());
-}
-
-#[test]
-fn readonly_route_auth_matches_basic_header() {
-    let auth = ReadonlyRouteAuth::from_config(Some(nodelite_proto::ReadonlyAuthConfig {
-        username: "viewer".to_string(),
-        password: "secret".to_string(),
-        enable_2fa: false,
-        totp_secret: None,
-    }));
-    let request = Request::builder()
-        .uri("/api/overview")
-        .header(header::AUTHORIZATION, "Basic dmlld2VyOnNlY3JldA==")
-        .body(Body::empty())
-        .expect("request should build");
-
-    assert!(auth.is_authorized(&request));
-}
-
-#[test]
-fn two_factor_session_cookie_must_be_server_issued() {
-    let sessions = TwoFactorSessions::new();
-    assert!(!sessions.is_authenticated("verified"));
-
-    let token = sessions
-        .create_authenticated()
-        .expect("session token should be generated");
-    assert!(sessions.is_authenticated(&token));
-    sessions.remove_authenticated(&token);
-    assert!(!sessions.is_authenticated(&token));
-}
-
-#[test]
-fn pending_session_invalidated_after_max_failed_attempts() {
-    let sessions = TwoFactorSessions::new();
-    let token = sessions
-        .create_pending()
-        .expect("pending session should be created");
-    assert!(sessions.pending_exists(&token));
-
-    // 前 N-1 次失败:pending 仍然有效。
-    for _ in 0..(crate::auth::TWO_FACTOR_MAX_FAILED_ATTEMPTS - 1) {
-        assert!(!sessions.record_failed_attempt(&token));
-        assert!(sessions.pending_exists(&token));
-    }
-
-    // 第 N 次失败:pending 必须被强制失效。
-    assert!(sessions.record_failed_attempt(&token));
-    assert!(!sessions.pending_exists(&token));
-
-    // 已经被失效的 token 再次记录失败时,应当也返回 true(等同已失效),
-    // 防止调用方因为找不到 pending 而漏掉"通知客户端清 cookie"的动作。
-    assert!(sessions.record_failed_attempt(&token));
-}
-
-#[test]
-fn totp_step_marked_used_blocks_replay() {
-    let sessions = TwoFactorSessions::new();
-    let step = 12345_u64;
-    let replay_retention =
-        std::time::Duration::from_secs(crate::auth::TWO_FACTOR_TOTP_REPLAY_RETENTION_SECS);
-    assert!(replay_retention >= std::time::Duration::from_secs(150));
-    assert!(!sessions.is_totp_step_used(step));
-    sessions.mark_totp_step_used(step);
-    assert!(sessions.is_totp_step_used(step));
-    // 不同 step 不会被误判
-    assert!(!sessions.is_totp_step_used(step + 1));
-    assert!(!sessions.is_totp_step_used(step - 1));
-}
-
-#[test]
-fn constant_time_compare_matches_only_identical_byte_slices() {
-    assert!(crate::auth::constant_time_compare_bytes(
-        b"abc123", b"abc123"
-    ));
-    assert!(!crate::auth::constant_time_compare_bytes(
-        b"abc123", b"abc124"
-    ));
-    assert!(!crate::auth::constant_time_compare_bytes(b"abc", b"abc1"));
-    assert!(!crate::auth::constant_time_compare_bytes(b"", b"a"));
-    assert!(crate::auth::constant_time_compare_bytes(b"", b""));
-}
-
-#[test]
-fn warns_for_remote_http_public_base_url() {
-    assert!(uses_insecure_remote_public_base_url(
-        "http://monitor.example.com",
-        SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::UNSPECIFIED, 8080)),
-    ));
-    assert!(uses_insecure_remote_public_base_url(
-        "http://203.0.113.10:8080",
-        SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 8080)),
-    ));
-}
-
-#[test]
-fn ignores_local_or_tls_public_base_url() {
-    assert!(!uses_insecure_remote_public_base_url(
-        "https://monitor.example.com",
-        SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::UNSPECIFIED, 8080)),
-    ));
-    assert!(!uses_insecure_remote_public_base_url(
-        "http://127.0.0.1:8080",
-        SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 8080)),
-    ));
-    assert!(!uses_insecure_remote_public_base_url(
-        "http://localhost:8080",
-        SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 8080)),
-    ));
-}
-
-#[test]
-fn server_readiness_tracks_dependency_health() {
-    let readiness = ServerReadiness::new(true);
-    assert!(readiness.is_ready());
-    assert_eq!(readiness.status_label(), "ok");
-
-    readiness.mark_registry_reload_healthy(false);
-    assert!(!readiness.is_ready());
-    assert_eq!(readiness.status_label(), "degraded");
-
-    readiness.mark_registry_reload_healthy(true);
-    readiness.mark_history_available(false);
-    assert!(!readiness.is_ready());
-    assert!(!readiness.history_available());
-}
-
-#[test]
-fn install_endpoints_disable_caching() {
-    let runtime = Runtime::new().expect("runtime should build");
-    runtime.block_on(async {
-        let script_response = install_agent_script().await;
-        assert_eq!(
-            script_response.headers().get(header::CACHE_CONTROL),
-            Some(&header::HeaderValue::from_static(
-                "no-store, no-cache, must-revalidate",
-            )),
-        );
-        assert_eq!(
-            script_response.headers().get(header::PRAGMA),
-            Some(&header::HeaderValue::from_static("no-cache")),
-        );
-
-        let unique = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("clock should be monotonic enough")
-            .as_nanos();
-        let temp_dir = std::env::temp_dir().join(format!("nodelite-bootstrap-cache-test-{unique}"));
-        std::fs::create_dir_all(&temp_dir).expect("temp dir should exist");
-        let registry_path = temp_dir.join("server.json");
-        let issued = issue_node(
-            &registry_path,
-            IssueNodeRequest {
-                node_id: "osaka-01".to_string(),
-                node_label: Some("Osaka 01".to_string()),
-                tags: Vec::new(),
-            },
-        )
-        .await
-        .expect("node should be issued");
-        let mut config = test_server_config(
-            SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 8080)),
-            "https://monitor.example.com".to_string(),
-            registry_path.clone(),
-            temp_dir.join("history.sqlite3"),
-            temp_dir.join("snapshot.json"),
-        );
-        config.readonly_auth = None;
-        config.ws = test_ws_config(32, 8);
-        config.stale_after_secs = 20;
-        config.ping_interval_secs = 10;
-        config.ignored_filesystems = Vec::new();
-        let state = AppState::test_fixture(config.into(), Arc::new(temp_dir.join("server.toml")))
-            .await
-            .expect("state fixture should build");
-        let request = Request::builder()
-            .uri("/install/bootstrap")
-            .header(
-                header::AUTHORIZATION,
-                format!("Bearer {}", issued.install_token),
-            )
-            .body(Body::empty())
-            .expect("request should build");
-        let peer_addr = SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 51234));
-        let bootstrap_response = install_bootstrap(
-            State(state),
-            ConnectInfo(peer_addr),
-            HeaderMap::new(),
-            request,
-        )
-        .await;
-        assert_eq!(bootstrap_response.status(), StatusCode::OK);
-        assert_eq!(
-            bootstrap_response.headers().get(header::CACHE_CONTROL),
-            Some(&header::HeaderValue::from_static(
-                "no-store, no-cache, must-revalidate",
-            )),
-        );
-        assert_eq!(
-            bootstrap_response.headers().get(header::PRAGMA),
-            Some(&header::HeaderValue::from_static("no-cache")),
-        );
-
-        let _ = std::fs::remove_dir_all(&temp_dir);
-    });
-}
-
-#[test]
-fn router_compresses_text_assets_but_not_webp() {
-    let runtime = Runtime::new().expect("runtime should build");
-    runtime.block_on(async {
-        let unique = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("clock should be monotonic enough")
-            .as_nanos();
-        let temp_dir = std::env::temp_dir().join(format!("nodelite-compression-test-{unique}"));
-        std::fs::create_dir_all(&temp_dir).expect("temp dir should exist");
-        let registry_path = temp_dir.join("server.json");
-        let mut config = test_server_config(
-            SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 8080)),
-            "https://monitor.example.com".to_string(),
-            registry_path,
-            temp_dir.join("history.sqlite3"),
-            temp_dir.join("snapshot.json"),
-        );
-        config.readonly_auth = None;
-        config.ws = test_ws_config(32, 8);
-        let state = AppState::test_fixture(config.into(), Arc::new(temp_dir.join("server.toml")))
-            .await
-            .expect("state fixture should build");
-        let app = crate::startup::build_router(state.clone());
-
-        for path in ["/", "/assets/ui-i18n.json", "/metrics"] {
-            let response = app
-                .clone()
-                .oneshot(
-                    Request::builder()
-                        .uri(path)
-                        .header(header::ACCEPT_ENCODING, "br, gzip")
-                        .body(Body::empty())
-                        .expect("request should build"),
-                )
-                .await
-                .expect("response should be produced");
-            assert_eq!(response.status(), StatusCode::OK, "{path}");
-            assert!(
-                response
-                    .headers()
-                    .get(header::CONTENT_ENCODING)
-                    .is_some_and(|value| value == "br" || value == "gzip"),
-                "{path} should be compressed",
-            );
-        }
-
-        let webp_response = app
-            .oneshot(
-                Request::builder()
-                    .uri("/assets/brand-logo-dark.webp")
-                    .header(header::ACCEPT_ENCODING, "br, gzip")
-                    .body(Body::empty())
-                    .expect("request should build"),
-            )
-            .await
-            .expect("response should be produced");
-        assert_eq!(webp_response.status(), StatusCode::OK);
-        assert_eq!(
-            webp_response.headers().get(header::CONTENT_TYPE),
-            Some(&HeaderValue::from_static("image/webp")),
-        );
-        assert!(
-            webp_response
-                .headers()
-                .get(header::CONTENT_ENCODING)
-                .is_none(),
-            "WebP assets should not be recompressed",
-        );
-
-        state.history.shutdown().await;
-        state.audit_log.shutdown().await;
-        let _ = std::fs::remove_dir_all(&temp_dir);
-    });
-}
-
-#[test]
-fn spa_history_mode_routes_serve_index_shell() {
-    let runtime = Runtime::new().expect("runtime should build");
-    runtime.block_on(async {
-        let unique = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .expect("clock should be monotonic enough")
-            .as_nanos();
-        let temp_dir = std::env::temp_dir().join(format!("nodelite-spa-routes-test-{unique}"));
-        std::fs::create_dir_all(&temp_dir).expect("temp dir should exist");
-        let registry_path = temp_dir.join("server.json");
-        let mut config = test_server_config(
-            SocketAddr::V4(SocketAddrV4::new(Ipv4Addr::LOCALHOST, 8080)),
-            "https://monitor.example.com".to_string(),
-            registry_path,
-            temp_dir.join("history.sqlite3"),
-            temp_dir.join("snapshot.json"),
-        );
-        config.readonly_auth = None;
-        config.ws = test_ws_config(32, 8);
-        let state = AppState::test_fixture(config.into(), Arc::new(temp_dir.join("server.toml")))
-            .await
-            .expect("state fixture should build");
-        let app = crate::startup::build_router(state.clone());
-
-        // Every history-mode route in web/src/router/index.ts must return the SPA
-        // shell so deep links / refresh boot Vue instead of 404ing on the backend.
-        for path in ["/", "/nodes/osaka-01", "/settings", "/account", "/alerts"] {
-            let response = app
-                .clone()
-                .oneshot(
-                    Request::builder()
-                        .uri(path)
-                        .body(Body::empty())
-                        .expect("request should build"),
-                )
-                .await
-                .expect("response should be produced");
-            assert_eq!(
-                response.status(),
-                StatusCode::OK,
-                "{path} should serve the SPA shell"
-            );
-            assert_eq!(
-                response.headers().get(header::CONTENT_TYPE),
-                Some(&HeaderValue::from_static("text/html; charset=utf-8")),
-                "{path} should return index.html",
-            );
-        }
-
-        state.history.shutdown().await;
-        state.audit_log.shutdown().await;
-        let _ = std::fs::remove_dir_all(&temp_dir);
-    });
-}
 
 #[test]
 fn protected_routes_attach_security_headers() {
